Home / Dashboard
In the dashboard home screen add - labels to each of them, make sure that it is user-friendly or readable for the user.
Improve the graph from the dashboard it should be user-friendly and readable. Define the numbers from the graph like what are the x axis names about and the y axis. Add also the details of each line (legend) from the dashboard.